+++ app/src/main/java/app/olus/ytmusic/autolauncher/YTAutoLauncherApp.kt
@@ -1,6 +1,7 @@
 package app.olus.ytmusic.autolauncher
 
 import android.app.Application
+import app.olus.ytmusic.autolauncher.util.AALogger
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
@@ -8,4 +9,8 @@ class YTMusicAutoLauncherApp : Application() {
     
     var sharedUrlToProcess: String? = null
 
+    override fun onCreate() {
+        super.onCreate()
+        AALogger.init(this)
+    }
 }